+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNSDDF1RSSS2JJWXX6R80BBD"
+        "01KNSHM16Q7Y41YRN7Z0RQ3G6C"
     )
     r"""The unique identifier of the data source"""