+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNSDDF1QYNCA4WT96A27YAN6"
+        "01KNSHM16PQ1JY0TX68KFW9S8N"
     )
     r"""The unique identifier of the data source"""