+++ packages/orq-rc/RELEASES.md
@@ -7019,3 +7019,13 @@ Based on:
 - [python v4.8.0-rc.1] packages/orq-rc
 ### Releases
 - [PyPI v4.8.0-rc.1] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.1 - packages/orq-rc
+
+## 2026-04-09 16:38:56
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.1 (2.879.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.8.0-rc.2] packages/orq-rc
+### Releases
+- [PyPI v4.8.0-rc.2] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.2 - packages/orq-rc